data-and-ai-jhb/oct/code/claude-2/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
azure-ai-projects~=1.0.0b11
azure-identity~=1.20.0
ansible-core~=2.17.0
websockets==14.1
